+++ packages/core/src/interactivity/TooltipDataComponent.tsx
@@ -1,8 +1,7 @@
 import { createElement, MouseEvent, useCallback, useState } from 'react'
 import { DataComponentProps, InteractivityProps } from './types'
-import { cloneProps, CssProps, SvgElementVariantProps, useDimensions, WithId } from '../general'
+import { CssProps, SvgElementVariantProps, useDimensions, WithId } from '../general'
 import { useTooltip } from '../tooltips'
-import { merge } from 'lodash'
 import { getEventXY } from './utils'
 
 export const TooltipDataComponent = <
@@ -35,7 +34,7 @@ export const TooltipDataComponent = <
             handleTooltip(event)
             handlers?.onMouseEnter?.(data, event)
             if (modifiers?.onMouseEnter) {
-                setComponentStyle(merge(cloneProps(style, false), modifiers.onMouseEnter))
+                setComponentStyle({ ...style, ...modifiers.onMouseEnter })
             }
         },
         [data, handleTooltip, handlers, modifiers, style, setComponentStyle]
@@ -45,7 +44,7 @@ export const TooltipDataComponent = <
             handleTooltip(event)
             handlers?.onMouseMove?.(data, event)
             if (modifiers?.onMouseMove) {
-                setComponentStyle(merge(cloneProps(style, false), modifiers.onMouseMove))
+                setComponentStyle({ ...style, ...modifiers.onMouseMove })
             }
         },
         [data, handleTooltip, handlers, modifiers, style, setComponentStyle]
@@ -55,7 +54,7 @@ export const TooltipDataComponent = <
             setTooltipData({})
             handlers?.onMouseLeave?.(data, event)
             if (modifiers?.onMouseLeave) {
-                setComponentStyle(merge(cloneProps(style, false), modifiers.onMouseLeave))
+                setComponentStyle({ ...style, ...modifiers.onMouseLeave })
                 // setKey forces the component to repaint
                 // (without setKey, React can skip painting and the modifiers may not appear)
                 setKey(key => key + 1)
@@ -67,7 +66,7 @@ export const TooltipDataComponent = <
         (event: MouseEvent) => {
             handlers?.onClick?.(data, event)
             if (modifiers?.onClick) {
-                setComponentStyle(merge(cloneProps(style, false), modifiers.onClick))
+                setComponentStyle({ ...style, ...modifiers.onClick })
             }
         },
         [data, handlers, modifiers, style, setComponentStyle]